Enable \n replacement on all platoforms + Move logging to a module
